Time: 3 ms (14.8%), Space: 9 MB (97.61%) - LeetHub

+++ 0008-string-to-integer-atoi/0008-string-to-integer-atoi.cpp
@@ -24,7 +24,7 @@ public:
             if(sign==-1 && -ans<INT_MIN) return INT_MIN;
             i++;
         }
-        return sign*ans;
+        return (int)(sign*ans);
 
        }
